refactor: move Strategy components to /routes/strategies
