+++ src/routes/strategies/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import type { PageData } from './$types';
-	import { PageHeading, StrategiesListing } from '$lib/components';
+	import { PageHeading } from '$lib/components';
+	import StrategyList from './StrategyList.svelte';
 
 	export let data: PageData;
 	$: strategies = data.strategies;
@@ -10,7 +11,7 @@
 	<PageHeading title="<h1>Strategies</h1>" subtitle="<p>Currently available strategies</p>" />
 
 	{#if strategies.length}
-		<StrategiesListing {strategies} />
+		<StrategyList {strategies} />
 	{:else}
 		<p>No strategies configured</p>
 	{/if}
