+++ src/routes/strategies/StrategyTile.svelte
